requirements.txt
boto3==1.14.4
botocore==1.17.4
docutils==0.15.2
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
